refactor: remove githubid and use userid in mapping
