--- BE/src/main/java/com/codesquad/issue04/domain/entity/Issue.java
@@ -1,27 +1,0 @@
-package com.codesquad.issue04.domain.entity;
-
-import lombok.*;
-
-import javax.persistence.*;
-import java.util.List;
-
-@Getter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Builder
-@Entity
-public class Issue {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String title;
-
-    @OneToMany(mappedBy = "issue", cascade = CascadeType.REMOVE)
-    private List<Comment> comments;
-
-    @ManyToMany(mappedBy = "issue", cascade = CascadeType.ALL)
-    private List<Label> labels;
-    private String githubId;
-}
