+++ BE/src/main/java/com/codesquad/issue04/domain/issue/Issue.java
@@ -1,0 +1,47 @@
+package com.codesquad.issue04.domain.issue;
+
+import com.codesquad.issue04.domain.label.Label;
+import com.codesquad.issue04.domain.milestone.Milestone;
+import com.codesquad.issue04.domain.user.User;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.*;
+
+import javax.persistence.*;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@ToString
+@Builder
+@Entity
+public class Issue {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String title;
+
+    @OneToMany(fetch = FetchType.EAGER, mappedBy = "issue", cascade = CascadeType.REMOVE)
+    private List<Comment> comments;
+
+    @ManyToMany(fetch = FetchType.EAGER)
+    @JoinTable(
+        name = "label_has_issue",
+        joinColumns = @JoinColumn(name = "issue_id"),
+        inverseJoinColumns = @JoinColumn(name = "label_id")
+    )
+    private Set<Label> labels;
+
+    @ManyToOne
+    @JoinColumn(foreignKey = @ForeignKey(name = "milestone_id"))
+    private Milestone milestone;
+
+    @ManyToOne
+    @JoinColumn(foreignKey = @ForeignKey(name = "user_id"))
+    private User user;
+}
